mlnx-ofa_kernel: Upgrade OFED driver version to 5.5-1.0.3.2
This upgrades the OFED driver related packages to the ones that are
included in MLNX_OFED_SRC-5.5-1.0.3.2.tgz. The package can be
found here: https://content.mellanox.com/ofed/MLNX_OFED-5.5-1.0.3.2
/MLNX_OFED_SRC-5.5-1.0.3.2.tgz. also removes libibverbs because
that is now included in rdma-core package, and adds mlnx-tools
package that mlnx-ofa_kernel depends on.

The new versions are:
mlnx-ofa_kernel-5.5-OFED.5.5.1.0.3.1
rdma-core-55mlnx37-1.55103
mlnx-tools-5.2.0-0.55103

Reason:
The NVIDIA Mellanox ConnectX-6 device support is required by Cloud
Platform. The required Mellanox drivers must be upgraded to the latest
version (5.4+) to support the latest features of this device.

Verification:
- Build success based on the commit(kernel: Add auxiliary bus support)
- Install onto a StarlingX system with two controller and two compute
  nodes with network adapters Mellanox's OFED. The network adapters
  of controllers are Mellanox Technologies MT27710 Family
  [ConnectX-4 Lx], the network adapters of computes are Mellanox
  Technologies MT27800 Family [ConnectX-5].
- The bootstrap procedure was successful, Check dmesg output to see
  all mlx5_core driver information are correct.
- RDMA/Infiniband over Ethernet functionalities of the Mellanox adapters
  were successfully tested using the Linux RDMA community's perftest
  package.

Note that we do not test it on the NVIDIA Mellanox ConnectX-6, because
the network adapter is not available in our lab yet. We will test it
once ConnectX-6 is ready.

This work was based on the following commit by Babak Sarashki.
https://review.opendev.org/c/starlingx/kernel/+/822660

Story: 2009878
Task: 44610
Depends-On: https://review.opendev.org/c/starlingx/tools/+/831338

Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>
Change-Id: Ie2eae9a09b7f5a0a1d01a91cfd0349237b4ccdbf