debian: Upgrade ice, i40e, iavf to bundle 26.8
This debianizes commit 40dd7abe86eb4f3ca827da4c44fba51920775a26

which upgrades the intel driver versions to the ones listed
in Intel driver bundle 26.8 .  The list can be found here:

https://downloadmirror.intel.com/709681/Release_Notes_26.8_v1.1.pdf

The new versions are:
ice-1.7.16
i40e-2.17.4
iavf-4.3.19

Of note is that the iavf driver builds and links against
the auxiliary bus driver, but does not package it.  At runtime
it uses the one packaged in the ice driver which is the same
version.  Beware of this dependency if upgrading either driver.
This behaviour is the same as in CentOS.  At this time,
a backport of the auxiliary bus driver to kernel 5.10 is
being investigated as the Mellanox driver also has need to
use it, and managing version dependencies is quickly going
to be an issue if it is not built into the kernel.

Verification:
- ice, i40e, iavf packages all build ok
- examine the internals of the resulting .deb packages to
  ensure expected .ko files are there
- run build-image and verify that an iso is successfully
  generated

Story: 2009130
Task: 44608

Change-Id: I0ad22762b1de316d5ae4631f05db9569bff037d1
Signed-off-by: Jim Somerville <jim.somerville@windriver.com>