mlnx-ofa_kernel: Add mlnx firmware tools
This integrates NVIDIA firmware tools mstflint-4.16.0-1.55103.
This package contains a burning tool and diagnostic tools for
Mellanox manufactured HCA/NIC cards. The following tests are
done on CX4 and CX5.

Verification:
- Install onto a StarlingX system with two controller and two compute
  nodes with network adapters Mellanox's OFED. The network adapters
  of controllers are Mellanox Technologies MT27710 Family
  [ConnectX-4 Lx], the network adapters of computes are Mellanox
  Technologies MT27800 Family [ConnectX-5].
- Use mstflint to query the firmware on the device.
- Use mstflint to verify firmware.
- Use mstflint to update the firmware on the device.
- Use mstconfig to query configurations.
- Use mstvpd to dump the on-card VPD.
- Use mstregdump to dump hardware registers from Mellanox hardware.

Note that we do not test it on the NVIDIA Mellanox ConnectX-6, because
the network adapter NVIDIA Mellanox ConnectX-6 is not available in our
lab yet. We will test it once ConnectX-6 is ready.

Story: 2009878
Task: 44636
Depends-On: https://review.opendev.org/c/starlingx/kernel/+/831461

Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>
Change-Id: Ia41cd9d8b787f4a06c25e4ec8d15e8988abbfed1