kernel-rt: Do not build bpftool and perf packages
Prior to this commit, the bpftool and perf packages would be built by
kernel-std and kernel-rt spec files separately, and only the packages
built from the kernel-std spec file would be included in the ISO image.

However, merely having the kernel-rt spec file build the bpftool and
perf packages results in conflicts when preparing StarlingX binary patch
files that include all kernel and kernel-related RPM packages. Hence,
this commit makes the kernel-rt.spec file not build the bpftool and perf
packages by default.

We should also note that work in the mainline kernel is ongoing
(https://lwn.net/Articles/812503/) to allow BPF and to work well with
the preempt-rt patchset after concerns were raised
(https://lwn.net/Articles/802884/). The 5.10 kernel does have many of
these changes in place, but to provide the best performance while
avoiding any possible remaining issues with BPF, it is best not to build
bpftool and perf.

Verification:
- After rt kernel build is complete, confirm that the bpftool and perf
  packages for rt kernel do not exist in the RPM package directory,
  using the following command, which does not return output:
    ls "${MY_BUILD_DIR}/rt/rpmbuild/RPMS/" | grep -e bpftool -e perf
- Confirm that an ISO image built with this patch is successfully
  installed into two VMs in All-in-One simplex mode, one with the
  standard profile and another with the low-latency profile.
- Confirm that the aforementioned VMs' Ansible bootstrap procedures pass
  as well.

Closes-Bug: #1964432

Change-Id: Ifda79d15707a0d36dddd0fd19b75656f9aa5437a
Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>
[mvb: Update commit message to address code review comments and test the
 commit further.]
Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>