centos_extra_downloads.lst: Add mstflint SRPM
The addition of mstflint to the centos_extra_downloads.lst was forgotten
with commit 006c2597033b ("mlnx-ofa_kernel: Add mlnx firmware tools").
This was also not noticed during the code review.

Not having mstflint listed in the centos_extra_downloads.lst file
results in the following build-srpms preparation failure, especially
when the source RPM is not manually placed in
"${MY_REPO}/stx/downloads":

  $ ${MY_REPO_ROOT_DIR}/stx-tools/toCOPY/generate-centos-repo.sh \
    /import/mirrors/starlingx
  # Output redacted

  $ ${MY_REPO_ROOT_DIR}/stx-tools/toCOPY/populate_downloads.sh \
    /import/mirrors/starlingx
  # Output redacted

  $ build-srpms --clean mstflint && build-srpms mstflint
  # Some output redacted
  b2: ERROR: build_dir (438): Invalid srpm path \
    'repo:stx/downloads/mstflint-4.16.0-1.55103.src.rpm', \
    evaluated as \
    '.../stx/downloads/mstflint-4.16.0-1.55103.src.rpm', \
    found in \
    '.../stx/kernel/userspace/mellanox/mstflint/centos/srpm_path'
  # Some output redacted

This commit resolves this issue by adding the missing entry to the
centos_extra_downloads.lst file.

Change-Id: Ib49d9af6b40e027876e39a98d696b2b54f67f6bd
Closes-Bug: 1964543
Fixes: 006c2597033b ("mlnx-ofa_kernel: Add mlnx firmware tools")
Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>