CentOS: Upgrade ice, i40e, iavf drivers to bundle 27.1
This upgrades the intel driver versions to the ones listed
in Intel driver bundle 27.1. The list can be found here:
https://downloadmirror.intel.com/727294/Release_Notes_27.1.pdf
Also upgrades the ddp firmware to ice_comms-1.3.35.0 from
ice_comm-1.3.31.0.

0001-ice_xsk-Avoid-dependency-on-napi_busy_loop-with-PREE.patch
has been dropped because the fix had been integrated to the
driver ice-1.8.3.

The new versions are:
ice-1.8.3
i40e-2.18.9
iavf-4.2.2

Reason:
Update ICE driver to support the Intel Logan Beach NIC (E810-CQDA2T),
and that dictates the upgrade of the iavf driver which in turn
dictates the upgrade of the i40e driver. We select the versions
listed in official Intel driver bundle 27.1 to keep the driver
versions compatible with each other. The ice_comm ddp firmware
upgrade to version 1.3.35.0 is also dictated by the bundle.

Verification:
- installs from iso succeed on servers with ice(Intel Ethernet
  Controller E810) and i40e hw(Intel Ethernet Controller X710)
- interfaces are up and pass packets
- create vfs, ensure that they are picked up by the new iavf
  driver and that the interface can come up and pass packets
- Check dmesg to see DDP package is loaded successfully and
  the version is 1.3.35.0.

Note that we do not test it on Intel Logan Beach NIC, because
the network adapter is not available in our lab yet. We will test it
once Intel Logan Beach NIC is ready.

Story: 2009952
Task: 44895
Depends-On: https://review.opendev.org/c/starlingx/tools/+/836728

Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>
Change-Id: I93dcf00fe443c4f86db2f8a65e159f9da05acc29