kernel-rt: Fix up Obsoletes logic
Since the update to the v5.10.99 kernel, the revisions of the standard
(std) and the preempt-rt (rt) kernels have been the same. This has
inadvertently caused the installer to install kernel-rt-headers instead
of kernel-headers even when installing a std/non-rt system.

This occurs because kernel-rt-headers provides and obsoletes
kernel-headers, and because kernel-rt-headers' revision is greater than
the revision of the actual kernel-headers package due to the ".rt"
string in the revision of kernel-rt-headers. The issue can be seen in
more concrete terms when looking at the package information below.
"rt.el7" is lexicographically larger/greater than "el7".
- kernel-headers-5.10.99-200.31.tis.el7.x86_64
- kernel-rt-headers-5.10.99-200.31.tis.rt.el7.x86_64

This in turn prevents the application of StarlingX binary patches that
include kernel packages, because the patching system's dnf invocation is
instructed to install/use kernel-headers at patch installation time, but
the kernel-rt-headers package is installed.

We have considered the following options to resolve this issue:

- (1) Remove the Obsoletes logic altogether.
- (2) Increment the std kernel's base revision number from 200 to 201,
  while keeping the rt kernel's base revision number at 200.
- (3) Only obsolete kernel-headers packages whose versions are less than
  the base version of the kernel-rt-headers (i.e., 5.10.99).

Implementing (1) appears to be the correct way, but has a larger
fall-out. During the rt build, libbnxt_re depends on kernel-rt-headers.
However, the mock/chroot file system already has kernel-headers-4.18,
which prevents the installation of kernel-rt-headers because of
conflicting files in kernel-rt-headers-5.10.99 and kernel-headers-4.18.
We could resolve this by disabling the rt build of the userspace
packages in the kernel packaging repository, such as rdma-core,
libbnxt_re and the like, but this increases the scope of the resolution
for the issue we have, so we did not opt for this approach, even though
it ought to be a planned improvement.

Option (2) ought to be a viable option as well. This option assumes that
the std and rt kernels will always have the same base version (i.e.
5.10.99). We have not tested this option.

Option (3) is the approach we take. This approach is similar to option
(2) in terms of the pre-requisite assumption, and the intent is to make
kernel-rt-headers only obsolete kernel-headers packages with versions
less than the base kernel version (i.e., kernel-headers < 5.10.99). This
allows the build system to install kernel-rt-headers to replace
kernel-headers-4.18 during the libbnxt_re build, and prevents the
installer from using kernel-rt-headers as a replacement for
kernel-headers.

While debugging this issue, we have also noticed that the following
condition in both kernel and kernel-rt spec files always evalutes to
true, so we remove this condition:
  %if "0%{?variant}"

This is because of the use of a string comparison with double quotes.
Looking at the CentOS and Fedora repositories, we see that this
condition is in CentOS 8 Stream as of this writing [1], but no longer is
in CentOS 9 Beta [2] and Fedora rawhide [3] distributions.

[1] https://git.centos.org/rpms/kernel-rt/raw/c8s/f/SPECS/kernel.spec
[2] https://git.centos.org/rpms/kernel-rt/raw/c9-beta/f/SPECS/kernel.spec
[3] https://src.fedoraproject.org/rpms/kernel/blob/rawhide/f/kernel.spec

Finally, please note that this patch does not attempt to allow StarlingX
binary patches to work on currently broken systems based on the master
branch. That is, there is no in-place upgrade path from a broken std
installation with kernel-rt-headers package to a std installation with
kernel-headers package. We consider this acceptable given that the
master branch is a development branch.

Verification:
- Clean all packages in the StarlingX kernel packaging repository with
  "build-pkgs --clean <pkg names>".
- Build an ISO image with this patch using a monolithic incremental
  build procedure.
- Install the ISO image to a virtual machine in All-in-One (AiO) simplex
  mode with the std profile. Confirm that kernel-headers is installed,
  and not kernel-rt-headers.
- To build a test patch, increase all StarlingX package revisions in the
  kernel repository by 10 by modifying build_srpm.data files with "sed".
  Forcibly build the resulting packages, and prepare a StarlingX patch
  with all packages in the kernel packaging repository. Ensure that the
  patch includes rt packages for lowlatency profile only and std
  packages for the standard profile only by using the package selection
  arguments to patch_build.sh as required. (Example:
  "--controller-worker-lowlatency kernel-rt-core...rpm" ).
- Confirm that the resulting StarlingX patch can be applied to the
  standard profile installation. Confirm that kernel-headers is upgraded
  with kernel-headers as expected.
- Install the ISO image to a virtual machine again, but this time in AiO
  simplex mode with "rt" profile. Confirm that kernel-rt-headers is
  installed.
- Confirm that the aforementioned StarlingX patch can be applied to the
  installation with the "rt" profile.

Closes-Bug: 1969124

Change-Id: If1968bd43b360b7ffac2e57548dbbc496b9966dc
Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>