Debian: Upgrade ice, i40e, iavf drivers to bundle 27.1
This upgrades the intel driver versions to the ones listed
in Intel driver bundle 27.1. The list can be found here:
https://downloadmirror.intel.com/727294/Release_Notes_27.1.pdf
Also upgrades the ddp firmware to ice_comms-1.3.35.0 from
ice_comm-1.3.31.0.

0001-ice_xsk-Avoid-dependency-on-napi_busy_loop-with-PREE.patch
has been dropped because the fix had been integrated to the
driver ice-1.8.3.

The new versions are:
ice-1.8.3
i40e-2.18.9
iavf-4.2.2

Reason:
Update ICE driver to support the Intel Logan Beach NIC (E810-CQDA2T),
and that dictates the upgrade of the iavf driver which in turn
dictates the upgrade of the i40e driver. We select the versions
listed in official Intel driver bundle 27.1 to keep the driver
versions compatible with each other. The ice_comm ddp firmware
upgrade to version 1.3.35.0 is also dictated by the bundle.

Verification:
- build of the modules succeed for std and rt.
- build the std iso and boot up with qemu.
- build the rt iso and boot up with qemu.

Story: 2009952
Task: 44924

Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>
Change-Id: I84bb19eba7a12d8fcc9c5f24c89b4c2c10f53207