Debian: Upgrade OFED driver version to 5.5-1.0.3.2
This upgrades the OFED driver related packages to the ones that are
included in MLNX_OFED_SRC-5.5-1.0.3.2.tgz. The package can be
found here: https://content.mellanox.com/ofed/MLNX_OFED-5.5-1.0.3.2
/MLNX_OFED_SRC-5.5-1.0.3.2.tgz. Also adds both the mlnx-tools package
that mlnx-ofa_kernel depends on, and the firmware tool mstflint.

In addition, this also fixes the build errors for the packages mstflint
and mlnx-tools, and adds postinst, prerm files for rt build in
mlnx-ofa_kernel that are copied from the original postinst and prerm
files during the upgrade process.

The new versions are:
mlnx-ofa_kernel-5.5.tgz
rdma-core-55mlnx37.tgz
mstflint-4.16.0-1.tar.gz
mlnx-tools-5.2.0.tar.gz

Reason:
The NVIDIA Mellanox ConnectX-6 device support is required by some users.
The required Mellanox drivers must be upgraded to the latest version
(5.4+) to support the latest features of this device.

Verification:
- Build module success for kernel-std/kernel-rt.
- Build package success for rdma-core, mstflint and mlnx-tools.
- Build out std iso and boot up with qemu success.
- Build out rt iso and boot up with qemu success.

Story: 2009878
Task: 44923

Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>
Change-Id: Icb6d62b483ac89220c135fa05cbe99516c4a231e