kernel-rt/dl_hook: Fix base commit identifier
During the update to the v5.10.99 preempt-rt-enabled kernel, the base
commit identifier was forgotten to be updated to the linux-yocto
repository's v5.10.99 commit identifier. This commit fixes that.

Test Plan:
 - PASS: Build kernel-rt.
 - PASS: Build the 7 oot kernel modules for kernel-rt.
 - PASS: Build the test iso for kernel-rt and modules and boot up
         on qemu.
Tests are done with this commit and commit
<kernel-std/kernel-rt: Drop Android patches> integrated.
Also make sure that the old version kernel source pkg is removed
from the mirrors/starlingx/sources when building.

Story: 2009221
Task: 45150
Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>
Signed-off-by: Li Zhou <li.zhou@windriver.com>
Change-Id: I29a060fbc352d0b90bf20532f56f0577e7629248