Debian: kernel-rt: add support for multi kernels
Adding an extra '-rt' to abiname can change '5.10.0-6-amd64'
in file names/paths to '5.10.0-6-rt-amd64'.
In this way, the same names/paths in kernel-rt and kernel-std
packages can be avoided.
This is what multi kernels feature need.

With the abiname change, the "5.10.0-6" in some kernel-rt packages'
names is changed to "5.10.0-6-rt".
e.g. linux-rt-headers-5.10.0-6-amd64 is changed to
linux-rt-headers-5.10.0-6-rt-amd64.
So change this package name in the control files of the oot
kernel modules to avoid build errors.

Test plan:
 [Test together with all the patches about supporting multi-kernel
  feature (kernel part).]
 - PASS: Build std kernel and its kernel modules OK.
         Build std image OK.
         Boot on qemu OK.
         Boot on target ProLiant ML350 Gen10 OK.
 - PASS: Do above tests for kernel-rt too.

Depends-On: https://review.opendev.org/c/starlingx/tools/+/842030
Story: 2009221
Task: 45380

Signed-off-by: Li Zhou <li.zhou@windriver.com>
Change-Id: Id620e7588b702ed4f621cae4c5206688371d2d3e