Debian: remove rt suffix of mlnx-ofed-kernel-utils
There is a common file (/etc/modules-load.d/mlnx.conf) in the
mlnx-ofed-kernel-modules packages for std and rt kernel, which
can't be identified by the path 5.10.0-6 vs. 5.10.0-6-rt between
std and rt packages. It is same for std and rt packages.

Because the multi kernel feature wants to install the oot kernel
drivers' packages for std and rt kernel into one rootfs, error will
be reported such as:
dpkg: error processing archive .../[KENREL MOUDLE FOR RT].deb
trying to overwrite '.../[COMMON FILE NAME]',
which is also in package [KERNEL MODULE FOR STD].

mlnx-ofed-kernel-utils contains the common files of this module and
isn't related with std/rt. So remove the suffix @KERNEL_TYPE@ from it,
and move "mlnx.conf" from mlnx-ofed-kernel-modules@KERNEL_TYPE@
to mlnx-ofed-kernel-utils. Make mlnx-ofed-kernel-utils work as the
common package of this kernel module.
The dependency will install mlnx-ofed-kernel-utils together with
the driver package mlnx-ofed-kernel-modules.

Test plan:
 - PASS: Check that there aren't common files in the driver packages
         for kernel-std and kernel-rt.

 [Test together with all the patches about supporting multi-kernel
  feature (kernel part).]
 - PASS: Build std kernel and its kernel modules OK.
         Build std image OK.
         Boot on qemu OK.
         Boot on target ProLiant ML350 Gen10 OK.
 - PASS: Do above tests for kernel-rt too.

Depends-On: https://review.opendev.org/c/starlingx/tools/+/842030
Story: 2009221
Task: 45380

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>
Change-Id: I38d13341ac82c305d52d50a07586e7083065f2bf