Debian: change abiname in rt's init control file
The control file in the debian folder downloaded is the one generated
by debian default build config. We set our own configs and adapt the
code in debian folder, so a new control file will be generated by the
gencontrol function to replace the initial one in the build process.
This works fine for building kernel new source.
But in our build system, the package's dsc file is created using
the initial control file before any build starts. Then dsc_depends
parses the dsc file and decides the order of packages' build.

The commit <Debian: kernel-rt: add support for multi kernels> changes
the abiname, which affects the pkgs' names of rt kernel.
Here replace "5.10.0-6" with "5.10.0-6-rt" in the initial control file
so that the right dsc file can be created according to the packages'
dependency listed in control before that build start.

Test plan:
 [Test together with all the patches about supporting multi-kernel
  feature (kernel part).]
 - PASS: Build std kernel and its kernel modules OK.
         Build std image OK.
         Boot on qemu OK.
         Boot on target ProLiant ML350 Gen10 OK.
 - PASS: Do above tests for kernel-rt too.
 - PASS: Build all the packages successfully on a new and clean
         project.

Depends-On: https://review.opendev.org/c/starlingx/tools/+/842030
Story: 2009221
Task: 45380

Signed-off-by: Li Zhou <li.zhou@windriver.com>
Change-Id: I81e120e8b7966f9fca92804bc9294575cce7e71b