debian: Change hardcoded NIC naming scheme
In [1] it was decided to hardcode the NIC naming scheme to v238.
Testing on more labs showed inconsistencies CentOS vs Debian on
current hardcoded scheme.

Switching to a custom scheme, which will allow more customization
to be implemented in systemd.

Patch 0011 was renamed to fix a typo.
0011-Debian-Harcode-net.naming-scheme-in-CONFIG_CMDLINE.patch
-> 0011-Debian-Hardcode-net.naming-scheme-in-CONFIG_CMDLINE.patch

Tests on AIO-SX:
PASS: new parameter is picked up

[1]: https://review.opendev.org/c/starlingx/kernel/+/841369

Depends-On: https://review.opendev.org/c/starlingx/kernel/+/841369
Depends-On: https://review.opendev.org/c/starlingx/integ/+/842337
Story: 2009964
Task: 45402
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: Idffff795d11f18c0c44da5d696484a0d7cd3cb2c