kmod-ice: Restore OS default DDP firmware
Problem:
Apparently there was some confusion about adding the
ICE COMMS DDP FW, we wanted to include it for
testing purposes, but did not really want it to be
the default loaded upon bootup.  This causes grief to
at least two versions of DPDK, namely 20.11 and 21.11
when using GPRS Tunneling Protocol (GTP).

Solution:
Restore the OS default DDP FW version which
ships with the ice driver, and make it load by
default at bootup.  We also continue to leave the
COMMS version in the package so it can be tried if desired.

Verification:
- Boot up on server that contains ice network hardware
- Confirm that ice nework interfaces are working
- Check that the stx-ice.pkg symlink points to the OS default
DDP firmware and not the COMMS version
- run "dmesg | grep -i DDP" and confirm that the driver loaded
the OS default DDP firmware and not the COMMS version

Partial-Bug: 1978883

Fixes: a3e3f7e ("Ice Driver: use new comms ddp firmware")
Change-Id: I14758dc49f5740b8f5482143f4dfd49febff2234
Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>