Enable mlx5 onboard device udev naming
This commit introduces a patch which prioritizes the udev
(re)naming rules for mlx5 controlled devices to take the
onboard device name (if it exists) over the slot/path name.

This is consistent with the naming order used by the StarlingX
installer to create the 70-persistent-net.rules file.

It is also consistent with the naming order in the 99-default.link
file.

Without this patch, there could be an inconsistency with the
70-persistent-net.rules first re-naming the device to its
slot/path name, and then being overridden by the Mellanox specific
82-net-setup-link.rules.

Testing:

- Ensure if a system has mlx onboard devices, that the interfaces
  are named according to the onboard name. (CentOS)
- Note: Since I have depended on a third party with access to a
  (CentOS) system with onboard MLX devices, the Debian portion of
  this patch has been confirmed to build successfully, but not
  functionally tested.

Closes-Bug: 1981831

Signed-off-by: Steven Webster <steven.webster@windriver.com>
Change-Id: I79f308b39debd8e5ffd131bc90262a7ab6345e41