Debian: kernel: Parallelize xz archive compression
Without this commit, a 12-CPU build VM used by the author takes between
9 and 10 minutes to compress the standard/non-preempt-rt kernel's
linux-source-5.10.tar.xz source archive.

This commit parallelizes the compression operation, and this reduces the
compression time to about 1 minute, saving about 8 minutes during the
build procedure.

During the preparation of this commit, reduction of the compression
ratio was also experimented with, but was not opted for due to the
increased need for disk space.

Some of the measurements that were acquired with the same 12-CPU build
VM can be seen in the table below:

			Archive size	Compression time
			(Bytes)
baseline		120687336	9~10 min
--threads=0		"		~55 sec
--threads=0 -3		142357736	~21 sec
--threads=0 -0		174808508	~ 7 sec

Verification
- "build-pkgs -c -b rt,std -p linux,linux-rt" rebuilds the kernel
  successfully, but takes a shorter duration of time.
- This commit has been in the author's local repository for more than a
  month, and no issues were observed during the kernel rebuilds carried
  out during the author's day-to-day work.

Closes-Bug: 1988442
Change-Id: Ib7c701f3f690d6120dc5aca1d0d42862151a2fc0
Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>