kernel: Add SPR-5G-ISA/AVX512-FP16 support
This commit adds Sapphire Rapids 5G Instruction Set Architecture
(SPR-5G-ISA) support to the CentOS-based StarlingX kernel. This involves
AVX512-FP16 instructions, but note that these instructions are not
directly used by the kernel.

The benefits for StarlingX users are the ability to enumerate CPUs'
AVX512-FP16 capabilities, and the ability to start KVM-based virtual
machines that can make use of the capabilities in question. (Please note
that supporting AVX512-FP16 with KVM virtual machines requires patching
StarlingX's qemu-kvm-ev package in addition to this commit, as well.)

The cherry-picked commits were acquired from the v5.11 kernel release,
and all of them applied cleanly. The only change to the patches involved
the third patch, which was modified to remove a reference to a CPU
feature (X86_FEATURE_VM_PAGE_FLUSH) support for which is not provided by
StarlingX's v5.10 kernel baseline.

Test plan:

- CentOS-based StarlingX
  - Standard and preempt-rt kernels and all out-of-tree kernel modules
    were successfully built using a monolithic build procedure.
  - An ISO image was successfully built with this change.
  - The changes were confirmed to not negatively affect installation and
    Ansible boot-strap procedures in All-in-One Simplex virtual machines
    using standard and low-latency profiles.
  - Using a Sapphire Rapids-based server in All-in-One Simplex
    configuration, the aforementioned ISO image was installed and
    Ansible-bootstrapped, and the enumeration of the "avx512_fp16" CPU
    feature in /proc/cpuinfo was verified with the low-latency and
    standard kernels.

- Debian-based StarlingX
  - An ISO image was successfully built (in an incremental manner) with
    this change.
  - The changes were confirmed to not negatively affect installation and
    Ansible boot-strap procedures in All-in-One Simplex virtual machines
    using standard and low-latency profiles. (Due to time constraints,
    Debian-based StarlingX tests were carried out with virtual machines
    only.)

Story: 2010247
Task: 46073

Change-Id: I430de20651b6c4a0aa0d854d295b1760cb7b889c
Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>