Debian: mstflint: Rename libibmad-devel to libibmad-dev
jenkins build failed with the error ": Depends: libibmad-devel
but it is not installable". As the error message indicates we
correct this error by renaming libibmad-devel to libibmad-dev.

We added the dependency libibmad-devel refer to mstflint.spec in
the first place, and it builded successfully. So I didn't double
check the package name. Actually I do not understand why the build
can pass most of time except the two times in jenkins build.

Testing:
- Build mstflint successfully.

Closes-Bug: 1988376

Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>
Change-Id: I61979b9352aba84cf22b887927446bc1bdba5331