rcu: Avoid running boost kthreads on isolated CPUs
Problem:
The rcub tasks are allowed to run on isolated CPUs.  This
is a two-fold problem.  The first is that these rcub tasks
steal precious yielded CPU time away from kworker
tasks that need to run on the isolated cpu.  The second
is that they interfere with nohz_full behavior.

Solution:
This is already solved for us upstream, so we simply
backport the patch here.  It affines the rcub tasks
away from any nohz_full cpus.  Note that even though
this issue only pertains to our rt kernel, we also
apply the patch to the std kernel for completeness.

Verification:
- Boot rt kernel without this patch applied, and check the
affinity of the rcub tasks.  They are set to run on any
cpu, even the nohz_full ones.
- Boot rt kernel with this patch included, and now see that
the nohz_full cpus are excluded from rcub affinity.

Closes-Bug: 1992201

Signed-off-by: Jim Somerville <jim.somerville@windriver.com>
Change-Id: I53cbfeddcd7c89b17bd7465134c02311caa7914f