Debian: add package initramfs-trigger for reusing initramfs file
The initramfs file is rather large and will appear to change frequently.
This will expand the size of a patch in Debian for no good reason as the
original initramfs can generally be reused. To solve this bloat
problem, we introduce a new empty package called initramfs-trigger. The
job of this package is to indicate to the Debian patch build system
whether or not to include the new initramfs. If this empty package has
a version change, then the Debian patch build system will include the
new initramfs, otherwise the old initramfs will still be used and not
included in the patch.

TEST PLAN:
PASS - `build-pkgs -a; build-image`. Package initramfs-trigger exist in
       rootfs, version 1.0.stx.1.

Story: 2008862
Task: 46687

Change-Id: Ie986b12e4b9d1830e799e5c0ebeca763de3cfcf1
Signed-off-by: Zhang Xiao <xiao.zhang@windriver.com>